requirements.txt
fastapi
uvicorn
faiss-cpu
pillow
numpy
flask
flask-cors
torch
torchvision
numpy
gunicorn